increment_factor=0.2
One leg seems to hit the ground hard. It is thus interpreted that the swing foot tends to go underground to cover the desired trajectory of the swing leg. The ground plane or foot plane is thus lifted up so that the increment takes the believed underground landing, above by a factor of 0.2 times the step height.